+++ src/Feature/DocumentLinkProviderInterface.php
@@ -2,12 +2,10 @@
 
 namespace Symfony\Lsp\Feature;
 
+use Symfony\Lsp\Protocol\DocumentRequest;
+
 interface DocumentLinkProviderInterface
 {
-    /**
-     * @param array<array-key, mixed> $params
-     *
-     * @return list<array<array-key, mixed>>|null
-     */
-    public function links(array $params): ?array;
+    /** @return list<array<array-key, mixed>> */
+    public function links(DocumentRequest $request): array;
 }